Add files via upload
This repository contains a comprehensive exploratory data analysis (EDA) project focused on understanding the risk of loan defaults for a bank. The goal of this project is to analyze customer data to identify factors that contribute to the risk of defaulting on a loan and to provide insights that can help the bank improve its lending strategies and reduce risk. Introduction : The Bank Defaulters Risk Analysis project explores factors affecting loan default risk. By analyzing a dataset of bank customers, the project seeks to uncover patterns and trends related to loan default, with the aim of improving risk management practices. The analysis involves data cleaning, visualization, and statistical testing to identify key factors associated with default risk. Objectives: To explore the dataset for insights into loan default factors. To identify demographic and financial features that are most predictive of default risk. To provide actionable recommendations for the bank’s risk management strategies.